Practice Set 4 Answers

+++ practice_set2/ps2_prog04.py
@@ -1,3 +1,3 @@
-num_1 = int(input("Enter the divedend: "))
-num_2 = int(input("Enter the divisor: "))
+num_1 = float(input("Enter the divedend: "))
+num_2 = float(input("Enter the divisor: "))
 print(f"The quotient of the numbers is {int(num_1 / num_2)}")
